refactor: AnniversaryList 위치 수정 (widgets->entities)
