+++ src/entities/anniversary/index.ts
@@ -1,4 +1,4 @@
-export { AnniversaryItem } from './ui/AnniversaryItem/AnniversaryItem'
+export { AnniversaryList } from './ui/AnniversaryList/AnniversaryList'
 export { type Anniversary } from './model/types'
 export { useAnniversariesInRange } from './hooks/useAnniversariesInRage'
 export { generateExampleGuideAnniversaries } from './lib/helpers'
