--- src/widgets/AnniversaryList/index.ts
@@ -1,1 +1,0 @@
-export { AnniversaryList } from './ui/AnniversaryList'
